+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## [3.6.14](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.13...v3.6.14) (2023-12-29)
+
+
+### Bug Fixes
+
+* export type CurrencyInputOnChangeValues and update examples ([#331](https://github.com/cchanxzy/react-currency-input-field/issues/331)) ([8ea2c47](https://github.com/cchanxzy/react-currency-input-field/commit/8ea2c47a786390e64cc2fa4253710be568721431))
+
+## [3.6.13](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.12...v3.6.13) (2023-12-28)
+
+
+### Bug Fixes
+
+* Update input to empty string when `userValue` is `undefined` ([#323](https://github.com/cchanxzy/react-currency-input-field/issues/323)) ([30c5fcc](https://github.com/cchanxzy/react-currency-input-field/commit/30c5fccb18e9d341f08e3bb1c05b7e1f808e229a))
+
+## [3.6.12](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.11...v3.6.12) (2023-11-07)
+
+
+### Bug Fixes
+
+* fixedDecimalValue issue [#292](https://github.com/cchanxzy/react-currency-input-field/issues/292) and handle fixedDecimalValue 0 ([#313](https://github.com/cchanxzy/react-currency-input-field/issues/313)) ([5be14cb](https://github.com/cchanxzy/react-currency-input-field/commit/5be14cb54128ebe314271001832d5756d3a60d60))
+
+## [3.6.11](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.10...v3.6.11) (2023-06-23)
+
+
+### Bug Fixes
+
+* decimal separator issues ([#294](https://github.com/cchanxzy/react-currency-input-field/issues/294)) ([95d9e0a](https://github.com/cchanxzy/react-currency-input-field/commit/95d9e0a3a43309b38107c3a590d6dcaed3eb3d97))
+
+## [3.6.10](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.9...v3.6.10) (2023-02-03)
+
+
+### Bug Fixes
+
+* closes Cursor jumping issue if the ref passed isn't a ref object [#247](https://github.com/cchanxzy/react-currency-input-field/issues/247) ([#276](https://github.com/cchanxzy/react-currency-input-field/issues/276)) ([ef4e9e4](https://github.com/cchanxzy/react-currency-input-field/commit/ef4e9e442ba6dd693163fe978308e2c4d65c2731))
+
 ## [3.6.9](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.8...v3.6.9) (2022-10-25)
 
 ## [3.6.8](https://github.com/cchanxzy/react-currency-input-field/compare/v3.6.7...v3.6.8) (2022-10-20)